+++ tests/test_cv.py
@@ -1,10 +1,5 @@
-import sys
-import numpy as np
-import pandas as pd
 import pytest
-import sklearn
 import sklearn.metrics
-from sklearn.datasets import load_breast_cancer
 from sklearn.datasets import fetch_openml
 from sklearn.model_selection import train_test_split
 
@@ -33,7 +28,5 @@ def get_data():
 def test_cv_score(get_data):
     data = get_data
     model = LightGBMClassifier(databunch=data, random_state=RANDOM_SEED)
-    score, score_std = model.cross_val(print_metric=True)
-    assert score is not None
     score, score_std = model.cross_val_score(print_metric=True)
     assert score is not None